feat: dry-run fee estimation without funded inputs (#2097)
## Summary

- **Indexer dry-run no longer requires funded inputs.** A `dry_run` flag
is plumbed from `TariTransactionProcessor::new` through `FeeModule` into
`FeeState`. The two `InsufficientFeesPaid` abort gates
(`checkpoint_fee_intent` and `StateTracker::finalize`) skip rejection
when set; metering still runs, so `cost_breakdown` carries the true
cost. Validator (bootstrap) hardcodes `false`; the indexer's
`DryRunTransactionProcessor` hardcodes `true`. Mempool already rejects
dry-run transactions, so the flag can't reach consensus.
- **Wallet daemon dry-run handlers return a real fee estimate.**
`validators.claim_fees` and `nfts.transfer` previously returned
`fee_receipt.total_fees_paid()`, which is clamped to `max_fee`. Since
callers pass a placeholder `max_fee` (often 1) on dry-run, the response
was a "fee equals max_fee" tautology. Both now return `required_fees()`
(= `total_fees_charged + 1`), independent of the cap.
- **Wallet sqlite u64-as-i64 wraparound fixes** (surfaced while
exercising the estimate flow). `confidential_outputs.value` /
`stealth_outputs.value` are u64 amounts stored in signed `BigInt`
columns; values ≥ 2^63 wrap to negative.
- `*_lock_smallest_amount` ordered by `value.asc()`, so wrapped rows
sorted first and "smallest" returned the largest output. Now partitions
on `(value < 0)` first to recover u64 order.
- `*_get_unspent_balance` used SQL `SUM(value)` and panicked at
`BigDecimal::expect("overflow")` on a negative total. Now loads rows and
aggregates as u128 via `i64 as u64 as u128` reinterpret.
- **TS bindings + web UI** updated to surface the new dry-run estimate
fields and use `bigint` where appropriate.

## Test plan

- [ ] Indexer: submit a dry-run transaction from an account with
**zero** input fees; confirm the response includes a non-rejected
`cost_breakdown` and the per-source fee components.
- [ ] Wallet: call `validators.claim_fees` with `dry_run: true, max_fee:
1`; confirm the response `fee` reflects the true cost (e.g. ~136 uTARI)
instead of `1`.
- [ ] Wallet: call `transfer_nft` with `dry_run: true, max_fee: 1`;
confirm `fee` is the uncapped estimate and `fee_refunded` doesn't
underflow.
- [ ] Wallet: confirm a confidential transfer still picks the smallest
unspent output (ordering regression check) — manual or via an existing
integration test in `crates/wallet/storage_sqlite/`.
- [ ] Wallet: confirm balance queries for accounts with wrapped values
don't panic. Optional — a synthetic row with `value >= 2^63` would
exercise this; not expected in mainnet/testnet today.
- [ ] Web UI: dry-run flows for ClaimFees, ClaimBurn, PublishTemplate,
SendNft, SendMoney show a sensible suggested fee.
- [ ] Validator node: confirm dry-run flag stays `false` (sanity, since
mempool rejects dry-run anyway).

---------

+++ clients/validator_node_client/src/types.rs
@@ -285,7 +285,6 @@ pub struct GetBlocksResponse {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[cfg_attr(feature = "ts", derive(ts_rs::TS), ts(export, export_to = "validator-node-client/"))]
 pub struct GetBlocksCountResponse {
-    #[cfg_attr(feature = "ts", ts(type = "number"))]
     pub count: u64,
 }
 
@@ -296,7 +295,6 @@ pub struct GetBlocksCountResponse {
     ts(export, export_to = "validator-node-client/", rename = "VNLogEntry")
 )]
 pub struct LogEntry {
-    #[cfg_attr(feature = "ts", ts(type = "number"))]
     pub timestamp: u64,
     pub message: String,
     pub level: LogLevel,
@@ -342,7 +340,6 @@ pub struct GetNetworkCommitteeResponse {
     ts(export, export_to = "validator-node-client/", rename = "VNCommitteeShardInfo")
 )]
 pub struct CommitteeShardInfo {
-    #[cfg_attr(feature = "ts", ts(type = "number"))]
     pub shard: Shard,
     pub substate_address_range: RangeInclusive<SubstateAddress>,
     pub validators: Vec<ValidatorNode>,
@@ -462,7 +459,6 @@ pub struct GetCommsStatsResponse {
 #[cfg_attr(feature = "ts", derive(ts_rs::TS), ts(export, export_to = "validator-node-client/"))]
 pub struct GetEpochManagerStatsResponse {
     pub current_epoch: Epoch,
-    #[cfg_attr(feature = "ts", ts(type = "number"))]
     pub current_block_height: u64,
     #[cfg_attr(feature = "ts", ts(type = "string"))]
     pub current_block_hash: FixedHash,
@@ -482,10 +478,8 @@ pub struct GetBlockRequest {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[cfg_attr(feature = "ts", derive(ts_rs::TS), ts(export, export_to = "validator-node-client/"))]
 pub struct GetBlocksRequest {
-    #[cfg_attr(feature = "ts", ts(type = "number"))]
-    pub limit: u64,
-    #[cfg_attr(feature = "ts", ts(type = "number"))]
-    pub offset: u64,
+    pub limit: u32,
+    pub offset: u32,
     pub ordering_index: Option<usize>,
     pub ordering: Option<Ordering>,
     pub filter_index: Option<usize>,
